backend-fastapi/requirements.txt
fastapi==0.120.4
uvicorn[standard]==0.38.0
sqlmodel==0.0.24
SQLAlchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
pydantic==2.12.3
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1


